backend/requirements.txt
# DNAI Backend Requirements
# Food Recognition & Calorie Tracking System

# Core Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Authentication & Security
PyJWT>=2.8.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0

# AI & ML Dependencies
torch>=2.0.0
transformers>=4.35.0
Pillow>=10.0.0

# Vector Database - FAISS
faiss-cpu>=1.7.4
numpy>=1.24.0

# LangChain & Gemini Integration
langchain>=0.1.0
langchain-google-genai>=0.0.5
google-generativeai>=0.3.0

# Environment Variables
python-dotenv>=1.0.0

# Data Validation
pydantic>=2.0.0

# HTTP Client (for health checks)
httpx>=0.25.0
